requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.5
python-dotenv==1.0.1
cryptography==43.0.0
